Extract user interaction logic into Ui class
