+++ src/main/java/Duke.java
@@ -12,7 +12,6 @@ import java.time.format.DateTimeFormatterBuilder;
 import java.time.format.DateTimeParseException;
 import java.time.temporal.ChronoField;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -21,22 +20,10 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class Duke {
-    private static final String TEXT_LOGO = " ____        _\n"
-            + "|  _ \\ _   _| | _____\n"
-            + "| | | | | | | |/ / _ \\\n"
-            + "| |_| | |_| |   <  __/\n"
-            + "|____/ \\__,_|_|\\_\\___|";
-    private static final String TEXT_GREETING = "Hello! I'm Duke\n" + "What can I do for you?";
-    private static final String TEXT_GOODBYE = "Bye. Hope to see you again soon!";
-    private static final String TEXT_DIVIDER = "____________________________________________________________";
-    private static final String TEXT_ACKNOWLEDGE_LIST = "Here are the tasks in your list:";
-    private static final String TEXT_ACKNOWLEDGE_MARK = "Nice! I've marked this task as done:";
-    private static final String TEXT_ACKNOWLEDGE_UNMARK = "OK, I've marked this task as not done yet:";
-    private static final String TEXT_ACKNOWLEDGE_DELETE = "Noted. I've removed this task:";
-    private static final String TEXT_ACKNOWLEDGE_TASK = "Got it. I've added this task:";
     private static final String SAVE_DIR = "data";
     private static final String FILE_NAME = SAVE_DIR + "/duke.txt";
 
+    private final Ui ui = new Ui();
     private final List<Task> tasks = new ArrayList<>();
     private boolean shouldExit = false;
 
@@ -46,7 +33,7 @@ public class Duke {
 
     public void run() {
         Scanner scanner = new Scanner(System.in);
-        greet();
+        ui.greet();
         loadTasks();
 
         while (!shouldExit) {
@@ -73,10 +60,7 @@ public class Duke {
                 processInput(commandType, paramMap);
                 saveTasks();
             } catch (DukeException e) {
-                printDivider();
-                printTabbed(e.toString(), 1);
-                printDivider();
-                System.out.println();
+                ui.showError(e.toString());
             }
         }
     }
@@ -84,11 +68,11 @@ public class Duke {
     private void processInput(CommandType commandType, Map<String, String> paramMap) {
         switch (commandType) {
         case EXIT:
-            sayGoodbye();
+            ui.sayGoodbye();
             shouldExit = true;
             break;
         case LIST:
-            listTasks();
+            ui.listTasks(tasks);
             break;
         case MARK_TASK:
             markTask(getParamAsInt(paramMap, "id") - 1);
@@ -200,82 +184,22 @@ public class Duke {
 
     private void markTask(int index) {
         tasks.get(index).setDone(true);
-        printDivider();
-        printTabbed(TEXT_ACKNOWLEDGE_MARK, 1);
-        printTabbed(tasks.get(index).toString(), 3);
-        printDivider();
-        System.out.println();
+        ui.acknowledgeMark(tasks.get(index));
     }
 
     private void unmarkTask(int index) {
         tasks.get(index).setDone(false);
-        printDivider();
-        printTabbed(TEXT_ACKNOWLEDGE_UNMARK, 1);
-        printTabbed(tasks.get(index).toString(), 3);
-        printDivider();
-        System.out.println();
+        ui.acknowledgeUnmark(tasks.get(index));
     }
 
     private void addTask(Task task) {
         tasks.add(task);
-        printDivider();
-        printTabbed(TEXT_ACKNOWLEDGE_TASK, 1);
-        printTabbed(task.toString(), 3);
-        printTabbed("Now you have " + tasks.size() + " tasks in the list.", 1);
-        printDivider();
-        System.out.println();
+        ui.acknowledgeAdd(task, tasks.size());
     }
 
     private void deleteTask(int index) {
         Task deleted = tasks.get(index);
         tasks.remove(index);
-        printDivider();
-        printTabbed(TEXT_ACKNOWLEDGE_DELETE, 1);
-        printTabbed(deleted.toString(), 3);
-        printTabbed("Now you have " + tasks.size() + " tasks in the list.", 1);
-        printDivider();
-        System.out.println();
-    }
-
-    private void listTasks() {
-        printDivider();
-        printTabbed(TEXT_ACKNOWLEDGE_LIST, 1);
-        for (int i = 0; i < tasks.size(); i++) {
-            String entry = (i + 1) + "." + tasks.get(i).toString();
-            printTabbed(entry, 1);
-        }
-
-        printDivider();
-        System.out.println();
-    }
-
-    private void greet() {
-        printDivider();
-        printTabbed(TEXT_LOGO, 1);
-        System.out.println();
-        printTabbed(TEXT_GREETING, 1);
-        printDivider();
-        System.out.println();
-    }
-
-    private void sayGoodbye() {
-        printDivider();
-        printTabbed(TEXT_GOODBYE, 1);
-        printDivider();
-        System.out.println();
-    }
-
-    private void printDivider() {
-        printTabbed(TEXT_DIVIDER, 0);
-    }
-
-    private void printTabbed(String s, int padding) {
-        String[] lines = s.split("\n");
-        char[] whiteSpace = new char[padding];
-        Arrays.fill(whiteSpace, ' ');
-
-        for (String line : lines) {
-            System.out.println('\t' + String.valueOf(whiteSpace) + line);
-        }
+        ui.acknowledgeDelete(deleted, tasks.size());
     }
 }
